Changed area values
Changed area values to be calculated and analysed as square meters, considering the actual size of the voxel,  more precise and intuitive

+++ PP_AI_Studies/Assets/Scripts/JSONReader.cs
@@ -62,9 +62,9 @@ public static class JSONReader
         {
             tenant.AssociateGrid(grid);
             //Tenant Area Preferences
-            tenant.AreaPreferences = new Dictionary<SpaceFunction, int[]>();
-            var areaWorkPref = tenant.AreaPrefWork_S.Split('_').Select(p => int.Parse(p)).ToArray();
-            var areaLeisurePref = tenant.AreaPrefLeisure_S.Split('_').Select(p => int.Parse(p)).ToArray();
+            tenant.AreaPreferences = new Dictionary<SpaceFunction, float[]>();
+            var areaWorkPref = tenant.AreaPrefWork_S.Split('_').Select(p => float.Parse(p)).ToArray();
+            var areaLeisurePref = tenant.AreaPrefLeisure_S.Split('_').Select(p => float.Parse(p)).ToArray();
             tenant.AreaPreferences.Add(SpaceFunction.Work, areaWorkPref);
             tenant.AreaPreferences.Add(SpaceFunction.Leisure, areaLeisurePref);
